:recycle: refactor: serviço de chamada a api separado por tipo (user e post)
